Refactor build gradle for local GUI testing
