+++ build.gradle
@@ -22,8 +22,26 @@ checkstyle {
 
 test {
     useJUnitPlatform()
-
     finalizedBy jacocoTestReport
+
+    systemProperty 'testfx.setup.timeout', '60000'
+}
+
+task guiTests(type: Test) {
+    useJUnitPlatform()
+    include 'systemtests/**'
+}
+
+task nonGuiTests(type: Test) {
+    useJUnitPlatform()
+    include 'spleetwaise/address/**'
+    include 'spleetwaise/common/**'
+    include 'spleetwaise/transaction/**'
+}
+
+guiTests.mustRunAfter nonGuiTests
+task allTests {
+    dependsOn nonGuiTests, guiTests
 }
 
 task coverage(type: JacocoReport) {
@@ -40,66 +58,6 @@ task coverage(type: JacocoReport) {
         xml.required = true
     }
 }
-
-task(guiTests)
-task(nonGuiTests)
-guiTests.dependsOn test
-nonGuiTests.dependsOn test
-
-task(allTests)
-allTests.dependsOn guiTests
-allTests.dependsOn nonGuiTests
-
-test {
-    systemProperty 'testfx.setup.timeout', '60000'
-
-    jacoco {
-        destinationFile = new File("${buildDir}/jacoco/test.exec")
-    }
-
-    doFirst {
-        boolean runGuiTests = gradle.taskGraph.hasTask(guiTests)
-        boolean runNonGuiTests = gradle.taskGraph.hasTask(nonGuiTests)
-
-        if (!runGuiTests && !runNonGuiTests) {
-            runGuiTests = true
-            runNonGuiTests = true
-        }
-
-        if (runNonGuiTests) {
-            test.include 'spleetwaise/address/**'
-            test.include 'spleetwaise/commons/**'
-            test.include 'spleetwaise/transaction/**'
-        }
-
-        if (runGuiTests) {
-            test.include 'spleetwaise/address/ui/**'
-            test.include 'spleetwaise/commons/ui/**'
-            test.include 'spleetwaise/transaction/ui/**'
-        }
-
-        if (!runGuiTests) {
-            test.exclude 'spleetwaise/address/ui/**'
-            test.exclude 'spleetwaise/commons/ui/**'
-            test.exclude 'spleetwaise/transaction/ui/**'
-        }
-    }
-}
-
-task headless {
-    doLast {
-        println 'Setting headless mode properties.'
-        test {
-            systemProperties = [
-                    'testfx.robot': 'glass',
-                    'testfx.headless': 'true',
-                    'prism.order': 'sw',
-                    'prism.text': 't2k',
-            ]
-        }
-    }
-}
-test.mustRunAfter headless
 
 dependencies {
     String jUnitVersion = '5.4.0'
@@ -127,7 +85,6 @@ dependencies {
 
     testImplementation "org.testfx:testfx-core:4.0.18"
     testImplementation "org.testfx:testfx-junit5:4.0.18"
-    testImplementation 'org.testfx:openjfx-monocle:17.0.10'
     testImplementation group: 'org.hamcrest', name: 'hamcrest', version: '2.1'
     testImplementation group: 'org.assertj', name: 'assertj-core', version: '3.13.2'
 }
@@ -136,4 +93,4 @@ shadowJar {
     archiveFileName = 'addressbook.jar'
 }
 
-defaultTasks 'clean', 'headless', 'allTests', 'coverage'
+defaultTasks 'clean', 'nonGuiTests', 'coverage'
